+++ persistance/src/main/java/de/failender/dgo/persistance/held/HeldRepositoryService.java
@@ -3,7 +3,7 @@ package de.failender.dgo.persistance.held;
 import java.util.List;
 
 public class HeldRepositoryService {
-    public static List<HeldEntity> findByUserId(Integer id) {
+    public static List<HeldEntity> findByUserId(Long id) {
         return HeldRepository.findByUserId(id);
     }
