+++ persistance/src/main/java/de/failender/dgo/persistance/user/UserRepository.java
@@ -1,12 +1,8 @@
 package de.failender.dgo.persistance.user;
 
-import de.failender.dgo.persistance.BaseRepository;
-import de.failender.dgo.persistance.HibernateUtil;
 import de.failender.ezql.EzqlConnector;
 import de.failender.ezql.queries.InsertQuery;
 import de.failender.ezql.queries.SelectQuery;
-import org.hibernate.Session;
-import org.hibernate.query.Query;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -22,7 +18,7 @@ class UserRepository {
 		return SelectQuery.Builder.selectAll(UserMapper.INSTANCE).build().execute();
 	}
 
-	public static UserEntity findById(Integer id) {
+	public static UserEntity findById(Long id) {
 		return firstOrNull(SelectQuery.Builder.selectAll(UserMapper.INSTANCE).where(UserMapper.ID, id).limit(1).build().execute());
 	}
 
@@ -30,7 +26,7 @@ class UserRepository {
 		return firstOrNull(SelectQuery.Builder.selectAll(UserMapper.INSTANCE).where(UserMapper.USER_NAME, name).limit(1).build().execute());
 	}
 
-	public static List<String> getPermissionsForUser(int id) {
+	public static List<String> getPermissionsForUser(Long id) {
 		try {
 
 			ResultSet rs = EzqlConnector.getConnection().createStatement().executeQuery("SELECT RIGHTS.NAME FROM USERS U INNER JOIN ROLES_TO_USER RTU ON RTU.USER_ID = U.ID INNER JOIN ROLES_TO_RIGHTS RTR ON RTR.ROLE_ID = RTU.ROLE_ID INNER JOIN RIGHTS ON RIGHTS.ID = RTR.RIGHT_ID WHERE U.ID = " + id);
